Renamed probability_map.py->map_array.py
