+++ openquake/hazardlib/map_array.py
